requirements.txt
advent-of-code-data
inquirer
python-dotenv
markdownify
watchdog
